Flatten to external/protocol
