+++ network/src/external/protocol/ping_pong_manager.rs
