+++ network/src/external/protocol/ping_pong_worker.rs
